Remove all the values when deleting a key
When deleting a key using `pthread_key_delete', delete all the values
associated to that key in all the threads available. Otherwise, the
key reuse in `pthread_key_create' can cause new keys to have thread
specific data of the previously used key with the same index.

Add a test for this case, which creates and deletes pairs of keys
checking that they have a NULL thread specific data after creation.

* sysdeps/hurd/pt-key-delete.c (pthread_key_delete): Remove all the
values of the key in all the threads.

* tests/Makefile (CHECK_SRC): Add test-17.c.
* tests/test-17.c: New file.