Set elf=yes if not defined already
glibc 2.16 dropped support for any binary format other than ELF, and the ELF variable has been removed too;
this causes a build failure when built as glibc addon with glibc >= 2.16, since it would not link to ld.so.
Defining ELF to 'yes' if not defined allows to retain compatibility with older glibc versions.

* Makefile [$(..) != ''] [$(elf) = ''] (elf): Define to 'yes'.